Tsuro/requirements.txt
aggdraw
Pillow
